Squares and colors fulling working. (Perhaps some bugs?)

+++ nBack/BackTypeResult+CoreDataProperties.swift
@@ -1,9 +1,9 @@
 import CoreData
 
 extension BackTypeResult {
-    @NSManaged public var matches: Int64
-    @NSManaged public var incorrect: Int64
-    @NSManaged public var correct: Int64
-    @NSManaged public var backType: Int64
+    @NSManaged public var matches: Int16
+    @NSManaged public var incorrect: Int16
+    @NSManaged public var correct: Int16
+    @NSManaged public var backType: Int16
     @NSManaged public var game: GameResult
 }
